requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-auth==2.38.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
multidict==6.3.2
prompt_toolkit==3.0.50
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyfcm==2.0.8
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-http-client==3.3.7
python-socketio==5.12.1
redis==5.2.1
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
twilio==9.5.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.19.0
